FileUpload/requirements.txt
SQLAlchemy==1.2.6
Flask_Migrate==2.1.1
Flask_SQLAlchemy==2.3.2
Flask==0.12.2
Werkzeug==0.14.1
short_url==1.2.2
alembic==0.9.9
Flask_Script==2.0.6
cropresize2==0.2.0
Pillow==5.1.0
python_magic==0.4.15
